+++ packages/docs/src/components/codesandbox/files/index.ts
@@ -1,14 +1,30 @@
-/**
- * Sandpack/CodeSandbox integration for nuqs demos
- * 
- * This module provides:
- * - Component source code for the Sandpack environment
- * - Demo templates showing nuqs usage
- * - File configurations for Sandpack
- */
+import type { SandpackFiles } from '@codesandbox/sandpack-react'
+import {
+  APP_COMPONENT,
+  INDEX_COMPONENT,
+  QUERY_SPY_COMPONENT,
+  QUERYSTRING_COMPONENT,
+} from './components'
 
 // Main exports - these are the primary public API
-export { SANDPACK_FILES } from './files'
+export const SANDPACK_FILES: SandpackFiles = {
+  '/App.tsx': {
+    code: APP_COMPONENT,
+    hidden: true,
+  },
+  '/QuerySpy.tsx': {
+    code: QUERY_SPY_COMPONENT,
+    hidden: true,
+  },
+  '/querystring.tsx': {
+    code: QUERYSTRING_COMPONENT,
+    hidden: true,
+  },
+  '/index.tsx': {
+    code: INDEX_COMPONENT,
+    hidden: true,
+  },
+}
 
 // Component exports - available if needed for customization
 export {
